Return a mediaUrl for story videos
The HTML structure of a story video is like this:

```
<video>
  <source src="...">
  <source src="...">
</video>
```

The video element has no src attribute, which is why the extension was
breaking. The source elements contain videos of the same format but
different resolutions. It's not possible to use any HTML method to know
which source element links to the highest resolution version of the
video, but it is observable that it's always the first source.

This approach is a bit blunt but seems to work.

Also added is a last-resort fallback on `currentSrc`, which returns the
src of the source element which happens to be used. It's not guaranteed
that it will return the highest resolution.