Fix download button
The download attribute which this extension depends on has a
restriction, that it only works on urls which are the same origin as the
page containing the link element.

Recently, either Instagram changed its asset host url, or Chrome began
respecting that rule when it didn't previously. As a result, the
download button began opening the image link within the UI iFrame,
rather than downloading the image.

This change fetches the image data and creates a local object url to the
image blob, which we use for our download button. The local object url
is always the same origin, and satisfies the download constraints.