Always download/copy video source, never thumbnail
Previously we would stop looking through the element stack once we hit
an image or video tag. On Instagram video posts, a thumbnail image
always sits above the video element when the video is paused. Clicking
on a video in this state would present the extension UI with the url
of the thumbnail image.

Now we continue to look through the element stack until we've hit the
body/html element, or reached a maximum depth of 15 (to prevent any
infinite loops.) As a result, we end up with a collection of a thumbnail
image and a video when a video is clicked. The code is flexible and
would work with a collection of any number of images or videos — it
always prefers videos, and will return the first of the stack if
present, and falls back on the first image in the stack otherwise.