Don't show extension when clicking slideshow arrows
Instagram's slideshow arrows sit above whatever media is currently being
displayed in the slideshow — this means that mediaAtPoint returns that
media's url when a slideshow arrow is clicked. The slideshow arrow also changes the media being displayed, causing the extension to use the previous media's url.

This adds slideshow arrows to the list of terminating elements, so that
if one sits above media, the media is never included in the elements
stack.

Because Instagram's classes may change, we have to detect the arrow with
a loose combination of signals. Originally I wanted to add all anchors
to the list of terminating elements, but the video play button is also
an anchor.