Display a loading state for the download button
This was never necessary previously because most assets on IG are
small and loaded in memory (once they're displayed.) Fetching an
image url that's already in memory results in an instant response.

IGTV videos can be 10 minutes long, and stream once played, so it can
take a few seconds for fetch to return a response.