Get largest src by bytesize rather than hardcoded width
Searching only for the src shown at a width of 1080px is very brittle.
Now, the extension fetches each src in a srcset and measures its
filesize, returning the url with the largest filesize.

This also updates the method of finding a media object's clipping
parent. Previously we searched through the element stack until we found
the first thing with overflow hidden. This could result in a false
positive, because sometimes the "hidden cover" layer above a media
object is display hidden, but *also* within the clipping parent, so we
don't want to use its dimensions. Now, the extension searches for
elements in the stack that are overflow hidden and shorter than the
media object.