Extension was failing due to calling .top on a null mediaRect
This error would break the app for the remainder of a session, until
browser reload.

This commit updates the render order in App.jsx so a null mediaRect
is never passed into a child component.

Additionally, loaderTimeout is cleared in a few more scenarios. In some
scenarios, the timeout was completing and causing App to render a child
component when there was no mediaRect. In other scenarios, the timeout
was completing after a mediaUrl had been set, causing a valid UI on the
screen to be replaced by an indefinite "loading" message.