Prevent incorrect rect from being set
Bug case: clicking on an image modal at an (x,y) coordinate that
included a profile page thumbnail in the background caused the
thumbnail's mediaRect to be passed via mediaAtPoint. Because thumbnails
are narrower than minimumImageWidth, no UI was being displayed for those
clicks, even when a valid image was clicked.

Instagram has added `overflow: hidden` to thumbnails. A hidden overflow
was the only signal that getClippingParentRect searched for when
determining if an element in the entire elementStack was a clipping
parent (see PR #32.) Therefore, when thumbnails were found in the
elementStack, they were assumed to be clipping parents, and their
mediaRect was passed instead of the actual asset's mediaRect.

This commit updates the clipping parent logic so that we check if the
media element is in fact a child of a potential clipping parent.