Show extension above DM image modal
Instagram has added DMs to the web. You can send images directly via DM,
and they appear in a new modal. This modal's z-index is 100, which is
the same z-index as this extension.

For the extension to appear above the modal regardless of DOM order, its
z-index must be higher. I've increased the z-index to a very high number
to prevent a similar bug from occurring in the near future.