Update the minTopValue to accommodate new stories design
The minTopValue is used to hide the extension UI while scrolling on the
main feed. Otherwise it appears above the sticky IG navigation. Stories
used to display smaller and centered and fell under this top value, but
now are displayed larger and closer to the top of the screen. The
extension was reading this top value and immediatly unmounting itself on
stories.

I updated the min value so the UI appears properly on stories, but it
now also appears over the sticky IG navigation on the feed page (until
it hits the new, smaller minTopValue.)

This should be followed up with a proper fix later.