Return video url on video pages
Does not return the correct url via the video modal on e.g. a profile
page with many videos or videos within slideshows.

Works by accessing the __additionalData and _sharedData objects on the
window via an injected script. A canonical video_url is stored there on
single video video pages.